backend/requirements.txt
fastapi
uvicorn
sqlmodel
google-generativeai
python-dotenv
bcrypt
python-jose[cryptography]
python-multipart
fastapi-mail
